apperror fix

+++ src/app/modules/Customer/customer.services.ts
@@ -1,11 +1,21 @@
+import AppError from "../../../shared/AppError";
 import { prisma } from "../../../shared/prismaClient";
-import { TCustomer } from "./customer.interface";
+import { TCustomer, TCustomerRest, TCustomerUpdate } from "./customer.interface";
+import httpStatus from "http-status"
 
 //Create Customer
 const createCustomer = async (payload: TCustomer): Promise<TCustomer> => {
   
-  if (!payload.name || !payload.email || !payload.phone) {
-    throw new Error("All fields are required.");
+  if (!payload.name || !payload.email || !payload.phone) {   
+   throw new AppError(httpStatus.NOT_FOUND, "All fields are required.");
+  } 
+  
+  const existingCustomer = await prisma.customer.findUnique({
+    where: { email: payload.email }
+  });
+ 
+  if (existingCustomer) {
+    throw new AppError(httpStatus.CONFLICT, "Email already exists. Please use a different email.");
   }
 
   const customerData = {
@@ -32,32 +42,59 @@ const getByIdFromDB = async (id: string): Promise<TCustomer | null>  => {
         where: {
           id: id
         }
-      })
-      if(!result){
-        throw new Error("Customer not found")
+      })   
+    if (!result) {
+      throw new AppError(httpStatus.NOT_FOUND, "Customer not found");
     }
     return result;
   };
 
   //Update from Database
-  const updateIntoDB = async (id: string, data: Partial<TCustomer>): Promise<TCustomer> => {  
+  const updateIntoDB = async (id: string, data: TCustomerUpdate): Promise<TCustomerUpdate> => {  
+    const existingCustomer = await prisma.customer.findUnique({
+      where: { id }
+    });
+  
+    if (!existingCustomer) {
+      throw new AppError(httpStatus.NOT_FOUND, "Customer not found");
+    }
     const result = await prisma.customer.update({
         where: {
           id: id
         },
-        data
+        data: {
+          name: data.name,
+          phone: data.phone,
+          email: undefined,
+        }
       })
-    return result;
+     
+      const { email, ...rest } = result;
+      return rest as TCustomerRest;
   };
 
 
   //delete customer from database
 const deleteFromDB = async (id: string): Promise<TCustomer | null>  => {  
+  if (!id) {
+    throw new AppError(httpStatus.BAD_REQUEST, "Customer ID is required");
+  }
+  
+  const customerExists = await prisma.customer.findUnique({
+    where: { id }
+  });
+
+  if (!customerExists) {
+    throw new AppError(httpStatus.NOT_FOUND, "Customer not found");
+  }
     const result = await prisma.customer.delete({
-        where: {
-          id: id
-        }
+      where: { id }
       })
+
+      if (!result) {
+        throw new AppError(httpStatus.NOT_FOUND, "Customer not found");
+      }
+      
     return result;
   };
